Introduce 'StorageBackend' abstraction
Files should be storable in many different places,
not just disk. A StorageBackend provision with pluggable
implementation provides for this.

Since the level of abstraction is now at 'string',
we will have to read the whole uploaded file into
memory. We will want to do JSON manipulation eventually
anyway, so this is ok.