Support pluggable naming conventions
We currently name things with sha256, but would
be really nice to name them with IPFS CIDs so we
can migrate to IPFS eventually - but without having
to tie ourselves to it.

Is it horrible to shell out for each upload? YES IT IS.
But CIDs aren't just hashes - they're chunked first,
and then the CID points to a list of chunks. So just
replicating the CID isn't gonna be enough.