Support uploading rendered R notebook files
Rendered R notebooks are .html files that can be reverse
engineered into .Rmd files. This is just reality, and we
should deal with it.

This actually adds support for *any* .html file, which is really
not what I want to do. Will need to put in safeguards later
so this doesn't become an arbitrary file host.

Ref https://github.com/notebook-sharing-space/ipynb.pub/issues/21