[WIP] Remove CC BY default licensing for content
In conversation with a lot of people, the project was
renamted from ipynb.pub to NotebookSharing.space,
with a focus on sharing notebooks with others - individual
or in a group - rather than *publishing*. The difference
has always seemed minor and cosmetic to me - but I think the
intent of the user needs to be considered. Sharing mostly
seems to be during collaboration, and as a quick 'look at this',
rather than something for perpetuity. The lack of any user
authentication adds to this - publication requires *intent*,
and *control* - and here we provide neither intentionally.
So the focus here is on in-progress work products rather
than 'final' work to be 'published'.

I've also been re-reading https://techautonomy.org, and found
it relevant. There's a general feeling of keeping the users
in control, and in this case the users are people who are
uploading notebooks. More specifically, there is

> Protect people's privacy and other rights by design

'Other rights' include copyright too. Copyright has been
used to fuck over so many people over and over (hello, Oracle!).
But it *feels* important to me that users *pick* this license,
rather than have it thrust upon them - at this level, at least.
GitHub should pick a default license for repositories at creation
- I think that's a good place for people to make an informed
choice about how they want their work to be used. Uploading
a single notebook doesn't seem to be such a place.